requirements.txt
# Python -> 3.9

jupyter 
tensorflow-gpu==2.9 
pandas 
numpy==1.26.4 
matplotlib 
imageio